fix dtype preservation

+++ scipy/optimize/_optimize.py
@@ -363,7 +363,7 @@ def rosen(x):
     xp = array_namespace(x)
     x = xp.asarray(x)
     r = xp.sum(100.0 * (x[1:] - x[:-1]**2.0)**2.0 + (1 - x[:-1])**2.0,
-               axis=0)
+               axis=0, dtype=x.dtype)
     return r
